Refactor output of phi on the particles
