+++ Source/Diagnostics/ParticleIO.cpp
@@ -235,3 +235,57 @@ MultiParticleContainer::WriteHeader (std::ostream& os) const
         allcontainers.at(i)->WriteHeader(os);
     }
 }
+
+void
+storePhiOnParticles ( PinnedMemoryParticleContainer& tmp,
+    int electrostatic_solver_id, bool is_full_diagnostic ) {
+
+    using PinnedParIter = typename PinnedMemoryParticleContainer::ParIterType;
+
+    WARPX_ALWAYS_ASSERT_WITH_MESSAGE(
+        (electrostatic_solver_id == ElectrostaticSolverAlgo::LabFrame) ||
+        (electrostatic_solver_id == ElectrostaticSolverAlgo::LabFrameElectroMagnetostatic),
+        "Output of the electrostatic potential (phi) on the particles was requested, "
+        "but this is only available for `warpx.do_electrostatic=labframe` or `labframe-electromagnetostatic`.");
+    // When this is not a full diagnostic, the particles are not written at the same physical time (i.e. PIC iteration)
+    // that they were collected. This happens for diagnostics that use buffering (e.g. BackTransformed, BoundaryScraping).
+    // Here `phi` is gathered at the iteration when particles are written (not collected) and is thus mismatched.
+    // To avoid confusion, we raise an error in this case.
+    WARPX_ALWAYS_ASSERT_WITH_MESSAGE(
+        is_full_diagnostic,
+        "Output of the electrostatic potential (phi) on the particles was requested, "
+        "but this is only available with `diag_type = Full`.");
+    tmp.AddRealComp("phi");
+    int const phi_index = tmp.getParticleComps().at("phi");
+    auto& warpx = WarpX::GetInstance();
+#ifdef AMREX_USE_OMP
+#pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
+#endif
+    for (int lev=0; lev<=warpx.finestLevel(); lev++) {
+        const amrex::Geometry& geom = warpx.Geom(lev);
+        auto plo = geom.ProbLoArray();
+        auto dxi = geom.InvCellSizeArray();
+        amrex::MultiFab const& phi = warpx.getField( FieldType::phi_fp, lev, 0 );
+
+        for (PinnedParIter pti(tmp, lev); pti.isValid(); ++pti) {
+
+            auto phi_grid = phi[pti].array();
+            const auto getPosition = GetParticlePosition<PIdx>(pti);
+            amrex::ParticleReal* phi_particle_arr = pti.GetStructOfArrays().GetRealData(phi_index).dataPtr();
+
+            // Loop over the particles and update their position
+            amrex::ParallelFor( pti.numParticles(),
+                [=] AMREX_GPU_DEVICE (long ip) {
+
+                    amrex::ParticleReal xp, yp, zp;
+                    getPosition(ip, xp, yp, zp);
+                    int i, j, k;
+                    amrex::Real W[AMREX_SPACEDIM][2];
+                    ablastr::particles::compute_weights(xp, yp, zp, plo, dxi, i, j, k, W);
+                    amrex::Real const phi_value  = ablastr::particles::interp_field_nodal(i, j, k, W, phi_grid);
+                    phi_particle_arr[ip] = phi_value;
+                }
+            );
+        }
+    }
+}
